Text changes
Made a new document - remembering to fix mouse as a first, still don't know how to tackle the rest of the box.
Also added a doll. Yes, I have a Jar Jar Binks doll. It is dirty and it uses batteries to say stuff. Unfortunately the battery connection pads have patina on them, so need to clean that as well.